[msvcrt] Update scanf family with a patch that also went to Wine (http://source.winehq.org/git/wine.git/?a=commit;h=2043035e1749e40e23853d751ac1410d39fc8d27): Calculate floating point decimals in internally used long double precision. Fixes bug #4092 (calculator imprecision)
svn path=/trunk/; revision=44304

+++ reactos/lib/sdk/crt/string/scanf.h
@@ -266,7 +266,7 @@ _FUNCTION_ {
 		    }
 		    /* handle decimals */
                     if (width!=0 && nch == '.') {
-                        float dec = 1;
+                        long double dec = 1;
                         nch = _GETC_(file);
 			if (width>0) width--;
                         while (width!=0 && (nch!=_EOF_) && _ISDIGIT_(nch)) {
